+++ src/main/java/scanner/Preferences.java
@@ -16,6 +16,7 @@ public class Preferences {
         prefs.put("-p", "554");
         prefs.put("-t", "10");
         prefs.put("-a", "5");
+        prefs.put("-bw", "2");
     }
 
     public static void configure(String[] values) {